+++ CHANGELOG.md
@@ -1,13 +1,13 @@
 # Arbimon Release Notes
 
-## v3.0.21 - Apr 15, 2021
+## v3.0.21 - Apr 16, 2021
 
 New features:
 
 - PI-520 Modify the Arbimon upload page
 
 Resolved issues:
-
+- PI-714 Fix timezone issue for calendar on the Visualizer page
 - PI-731 Bounding box is no longer visible in the Pattern Matching for Citizen Scientist users
 - PI-732 User unable to create PM template
 - Support for Core API v1.1.x changes